robust zero division calculator

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide(numerator, denominator):
     try:
-        result = int(numerator) / int(denominator)
+        result = float(numerator) / float(denominator)
     except ZeroDivisionError:
         return 'Error: Cannot divide by zero.'
     except ValueError:
